Listing protection added

+++ app/Http/Controllers/ListingController.php
@@ -60,6 +60,12 @@ class ListingController extends Controller
 
     //update listing data
     public function update(Request $request, Listing $listing){
+        
+        //makes sure logged in user is owner
+        if($listing->user_id != auth()->id){
+            abort(403, 'Unauthorized Action'); 
+        }
+
         $formFields = $request->validate([
             'title'=>'required',
             'company'=>'required',
@@ -83,8 +89,15 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->id){
+            abort(403, 'Unauthorized Action'); }
         $listing->delete();
         
         return redirect('/')->with('message','Listing Deleted');
     }
+
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
+    
 }
